third_party/WebKit/LayoutTests/platform/win/fast/events/offsetX-offsetY-svg-expected.txt

This is a testharness.js-based test.
PASS MouseEvent.offsetX/offsetY for an SVG element is always relative to the outermost <svg> CSS layout box 
Harness: the test ran to completion.
